all changes from local and server code merged.
added:
support for flag proxying from trusted relays
internal support for reverse ignore
SAYPRIVATEEX - acts exactly like SAYPRIVATE, but with EX so clients can show as action
check to make sure channels.xml isn't corrupted on server crash

fixed:
unicode error in client data able to crash server
invalid account lines in accounts.txt able to crash server
bugs from old code in SQLUsers
bug in loadargs
bug in GETIP
legacy bans

changed:
TESTLOGIN moved needs mod access instead of admin access
TESTLOGIN now responds with username's proper case and user's ID
OPENBATTLEFAILED type detection now trigger for non-integer ports
antispam is now more sane
cleaned up some code

known issues:
timestamps in LegacyUsers. need to check these, they're off by 1000x in an unknown direction.
BanHandler must be reloaded to get new bans (sqlalchemy is caching things somewhere)

+++ tasserver/LegacyChannels.py
@@ -1,6 +1,5 @@
 import xml.dom.minidom as minidom
 from xml.sax.saxutils import escape, unescape, quoteattr
-import binascii
 
 import os, re, codecs
 
@@ -79,7 +78,7 @@ class Parser:
 				topic = topics[name].decode('utf-8').encode('raw_unicode_escape') # chanserv writes double-encoded utf-8, this decodes it
 				
 			chan = {'owner':str(owner), 'key':str(channel.getAttribute('key')) or None, 'topic':topic or '', 'antispam':(str(channel.getAttribute('antispam')) == 'yes'), 'admins':chanops}
-			if chan[key] == '*': chan[key] = None
+			if chan['key'] == '*': chan['key'] = None
 			chans[name] = chan
 		
 		return chans
@@ -114,6 +113,6 @@ class Writer:
 		f.write('</channels>\n')
 		f.close()
 		
-		if os.path.exists('channels.xml'):
+		if os.path.exists('channels.xml') and os.path.exists('channels.xml.tmp'):
 			os.remove('channels.xml')
 		os.rename('channels.xml.tmp', 'channels.xml')
